+++ ginput/__init__.py
@@ -1,3 +1,3 @@
 # be sure to update in setup.py as well
 # please follow semantic versioning: https://semver.org/
-__version__ = '1.0.10'
+__version__ = '1.1.0'